requirements.txt
protobuf==3.19.4
grpcio==1.47.0
grpcio-tools==1.44.0
pandas~=2.0.0
requests~=2.28.1
urllib3~=1.26.15
pymongo~=4.3.3
pymysql~=1.0.3
pyyaml~=6.0
tqdm~=4.65.0
locust~=2.15.1
openpyxl~=3.1.2
